requirements.txt
asgiref==3.6.0
Django==4.1.6
Jinja2==3.1.2
MarkupSafe==2.1.2
sqlparse==0.4.3

djangorestframework~=3.14.0